Add temporary print for test

+++ src/verifier.cc
@@ -62,7 +62,7 @@ void CopyBitsOfMemoryToBuffer(uint8_t* dst,
 
   data >>= src_bit_offset;
   if (bits != 64)
-    data &= (static_cast<uint64_t>(1) << bits) - static_cast<uint64_t>(1);
+    data &= (1ULL << bits) - 1ULL;
   std::cout << __FILE__ << " " << __LINE__ << ": " << data << std::endl;
 
   std::memcpy(dst, &data, static_cast<size_t>((bits + 7) / 8));
